backend/requirements.txt
fastapi
uvicorn
langchain-groq
langgraph
sqlalchemy
pymysql
python-dotenv